+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 [Unreleased]: https://github.com/GemsTracker/gemstracker-library/compare/1.8.4...HEAD
+### Added
+- Compliance overview now provides a better export (#381)
 
 ## [1.8.5] - 2018-10-10
 [1.8.5]: https://github.com/GemsTracker/gemstracker-library/compare/1.8.4...1.8.5